Add the Context class
Right now, it can only be constructed and destructed. Small steps at a
time. There is error handling, but since the class doesn't really do
anything yet, and I'm not implementing the arguments, I can't trigger it
yet.

I suspect I will, at some point, need to change the deleter for the
unique_ptr to an custom delete function, but right, this is enough.