Rework DataNodes
Okay, so it seems the internals can be a lot simpler than I thought. Or
rather, very similar to the ones that were in libyang1. I think the main
problem with libyang1 C++ was that shared pointers were used for the
wrapper classes themselves. What it meant is that you could freely copy
these shared pointers and the deleters inside wouldn't know, or at least
wouldn't have too much control. Anyway, These wrappers do NOT have to be
allocated dynamically, they mostly serve as handles for the lyd_nodes
inside.

This makes everything a lot simpler. DataView class is no longer needed,
you can just use DataNode. DataNode classes are produced ONLY by the API
functions, so libyang knows about all the references. So, no problems
with dangling references and deletion.

The implementation uses a reference counter in the form of a
std::shared_ptr<Empty>. The object inside it doesn't do anything and the
deleter is the default one. The shared_ptr is only for it's reference
counting capabilities. Storing lyd_node directly inside the shared_ptr
is impossible: the data trees consist of multiple pointers and
shared_ptr only stores one pointer.