Add the DataNodeTerm class
This class wraps the lyd_node_term struct. It represents a leaf or a
leaf-list. Currently only supports a few operations.

The libyang1-cpp equivalent was called Data_Node_Leaf_List and you would
have to construct it by yourself. I don't want the user to construct any
Node objects, so DataNodeTerm is created by the DataNode::asTerm()
method. This works similar to the C-api - creating a DataNodeTerm class
is essentially just a typecast with an additional check.

Change-Id: I30cf5225209656d2673e2f17ca862124f17aeb37