Change include style
According to the cppreference, #include <x.hpp> searches "include
directories" (whatever that is), and #include "x.hpp" searches the
directory of the current file and then, if it doesn't find anything,
works as if it was the angled brackets version. However, all of this is
implementation defined.

The version with quotes works, but I feel like there is less control
over what directories are searched. I looked at how Boost.Spirit does it
and they use angled brackets, so that's what we're gonna too from now
on :).

Change-Id: I9c638fa7225af43b611d80d0585021fed48ccb78