Rework refcounting
Unlinking connecting and other manipulation of trees will require us to
remember all references to the current tree. This patch reworks the
refcounting to do that. The private member m_viewCount was also renamed
m_refs since that makes more sense now.

internal_empty was renamed to internal_refcount. The struct now contains
a member which holds all DataNode pointers to the tree.

Change-Id: I3e17186e34fe61452ab4dab5b15b015d151b7573