Implement Container::isPresence
There has been a discussion on how to implement the SchemaNode/DataNode
type system. There are three options:

1) Keep the current type system. Derived type like Container can be
retrieved by using an asX method. This method can possibly turn into a
template method as<X>.
2) Use inheritance and pointers. For example, if I'm retrieving a
SchemaNode which is a Container, I can dynamic_cast it to Container
immediately, without using any kind of special method.
3) Use std::variant in some way - I haven't done much research to this,
but it seems like a bad choice, because I need some sort of hierarchy (I
want the SchemaNode class as a generic node type, variant wouldn't have
a generic access without checking the type).

The question is whether it is preferable to use the C++ type system
instead of the current one. The main advantage of the C++ type system
and inheritance is that you don't reinvent the wheel. Although the
current type system isn't exactly complex, C++ already has a type
system, so it's worth considering using it. The main advantage is that
the user can use tools directly embedded into C++ like inheritance,
dynamic_cast and others.

On the other hand, it introduces complexity. All DataNode/SchemaNode
objects now have to be pointers. This makes everything a little bit more
complex. The classes don't need to be dynamically allocated, because the
classes won't use polymorphism (no virtual methods, you either don't
care about the type, so you use the base class' methods, or you know the
type, so you downcast to a derived class (like libyang::Container).
Another problem is using operator== on the nodes. Dereferencing them
before comparing is not a problem, but when you compare a std::variant
containing a shared_ptr, it is a little more difficult dereferencing.
You would either need to wrap the variant (which is undesirable IMHO)
or create a comparing function, which is meh IMHO. So from the
standpoint of the library developers, this only brings problems.

What about the user's standpoint?
Comparison of how to get the derived type:

For the current type system:
  node->asContainer(); or node->as<Container>();
If we were to change the system:
  dynamic_cast<Container>(node);

Okay, so this isn't so different. But what about chaining more methods?

Current type system:
  node->findPath("/mod:myLeaf")->schema()->as<Leaf>()->leafType();
With dynamic_cast:
  dynamic_cast<Leaf>(node->findPath("/mod:myLeaf")->schema())->leafType();

IMHO the first version is so much clearer, because you can read it from
left to right: you find a path, get the SchemaNode node for it, get a
Leaf from that and get the type of the leaf value. The dynamic_cast cast
version is different, because you need to wrap part of the expression in
dynamic_cast and only then use ->leafType().

Another argument is that users might not know that there is an as<X>
method: this is true, you would have to read the documentation for the
library to know that. But it's the same with dynamic_cast: you also have
to read the documentation (or the library implementation) to know that
SchemaNode/DataNode can actually be downcast to something more specific.
Also, if the user is using some kind of IDE, the as<> method might be
autocompleted. dynamic_cast won't be.

Another thing: what if the user just doesn't care about the specific
types like Container or Leaf? Then we're doing extra work, because the
user won't care about downcasting (but I don't know whether it would be
a noticeable performance issue).

I think the current system is better both from the developer's
standpoint and from the user's standpoint.

Change-Id: Ib2a72ec63711d09269ec7de699bc66caf8655b50