Implement DFS iterator
The old libyang1-cpp implemented this in a kind of weird way: it created
a vector of DataNode classes and returned that to the user. This always
seemed kind of weird to me, because DataNode::tree_dfs seemed like it
would return something iterable (something that .begin and .end), but
not necessarily a collection (like vector).

It turns out that wrapping the LYD_TREE_DFS_BEGIN isn't a simple task.
The macro itself is a for loop which defines some control variables.
Then there is the LYD_TREE_DFS_END macro that controls the iteration.
The problem with this macro is that it's really a "for-each" loop, that
gives you a variable. However, I don't want a loop. I need "something
that gives me the next node in the DFS algorithm". In other words, I
need to break from macro loop, return `elem` to the user, and when the
user calls operator++ on my iterator class, resume the loop. This is
however not possible, because of how the macros are implemented - you
can use `break`, but you can't resume again (AFAIK).

There a lot of solutions for this:
1) Write my own DFS logic - this way I don't need to wrap the macro and
all my problems are solved. Downside is I write something that's been
already written by the C-api.
1.5) Try to reverse engineer the macro, use clever `break`s here and
there, maybe expand it to actual code and save its control variables -
this way I could use the original code. However, it seems pretty messy.
It's more like a hack than a solution to me. And I'm not sure if it's
even a real solution (not sure if it's possible).
2) Use the libyang1-cpp solution - just complete the for-each loop and
return a std::vector<DataNode>. This is the simplest solution requires
the least work and works pretty well. It kinda goes against my intuition
in that it returns a collection and not an iterator. A tweak would be to
wrap the vector in a custom iterator class. But the vector would still
be there. Another downside is that you ALWAYS construct all of the
DataNode instances. This might not be so bad, since the classes are
small, but it's still an issue.
3) Create an iterator class that holds std::vector<lyd_node*> and create
DataNode instances and demand, when the user uses operator*. This is a
bit cleaner solution. Not sure how much I gain from this. If the user
does actually want all the DataNode instances (which they probably
will a lot of the time), then it's just a useless optimization.
4) Use C++20 coroutines to `pause` the macro for-each in the middle of
each iteration. This is a bold solution. It would mean to figure out how
coroutines work (but it's not like it's such complex thing, especially
for this use-case). In my small research, I found out that I would need
to create a `generator`. This solution is pretty cool I would say.
However, the problem with C++20 coroutines is that even though C++20
supports them, there isn't much library support for them. That means
that I would probably need to import a 3rd-party library like "cppcoro"
(or implement the stuff on my own, which is probably not a good idea,
because I'm not that big of a C++ mastermind).

After a discussion with Honza and Tomas, I have chosen version number 1.
The solution is the cleanest, I don't create unneeded objects. The only
downside is that I don't reuse the C macro, which is fine. It's not like
the DFS algorithm will change.

The algorithm inside DfsIterator is basically a copy of the
LYD_TREE_DFS_BEGIN and LYD_TREE_DFS_END macros. There are some
differences:
1) The algorithm is now an actual iterator, not a for-loop. The user can
use operator++ to advance the iteration and operator* to retrieve the
element.
2) Support for `LYD_TREE_DFS_continue` is removed. There is no need to
support this, because there's no loop inside the algorithm logic. The
user can use `continue` in their own loop as they would when iterating
over any other container.

The code also tracks the validity of collections and iterators. When a
subtree associated with a collection or an iterator gets unlinked, all
collections and iterators are invalidated. If a collection goes out of
scope before an iterator does, the iterator is also invalidated.

Change-Id: If0fbd8d3574408444c4ef8bebc276c30be525a67