Refactor DfsCollection to a separate file
Support for SchemaNode DFS iteration is coming next, so I need to
refactor the classes.

There were some problems with the DataNodeProxy class having a
forward-declared member. I solved this by making it a template class.
I don't know about another way to solve this, but I'm making all of the
classes templates anyway, so it shouldn't matter too much.

Change-Id: Ife60885c433eadc693cbf2fb16f3158457d13ded