Add SchemaNode::child
The method could probably be named `firstChild`, but I'll use this name,
because that's what the C api uses.

Side note: at first I thought implementing this method would be a PITS,
but thankfully the C-api includes a function (lysc_node_child), that
does the correct thing across all the schema node types.

Change-Id: I7cfbb3e6bc1bd205ef789019807f75278ca684df