Add Type::name
libyang has two types of schema nodes: compiled and parsed. The parsed
version contains all of the info from the schema like names of typedefs
and description, while the compiled version only contains resolved
information, so only basic info about the type is available.

The only way to get the parsed info is to use the
ContextOptions::SetPrivParsed flag when creating the context. Because of
this, the new Type::name method only works if the flag was used.
There are a few ways to disable the method:

1) purely runtime solution - check whether the context is created with
   flag and Type::name throws, if the flag wasn't used.
2) make the flag a compile time variable - make it so the Context class
   knows whether the user used the flag in compile time. This means,
   that I could completely remove the methods that would require the
   flag. The downside is that esentially ALL the classes would now have
   to be templates, because they would have to remember the flag.
3) just use the flag by default, even if the user didn't specify it and
   make the parsed info available always - Always using the flag does
   mean that the user loses the ability to save data into the `priv`
   field of the nodes. I'm not sure how useful is this in C++. Also, if
   the Context isn't created by us, but for example `libnetconf`, then
   this solution fails.

I discussed this with Tomas and found out that the best solution is to
use the runtime version (number 1). These are the reasons:
1) It is the easiest one to implement - throwing an exception is three
   lines of code. The exception will clearly say what the users should
   do if they want to use the method.
2) It doesn't make everything a template - I'm not sure if I would
   actually have to make everything a template, but I would definitely
   need at least some.
3) IMHO, having the method exist does bring some value - otherwise, the
   users might not know, that it could exist with some sort of Context
   flag.

There is one small disadvantage: the Type class now ALWAYS has a new
member variable (lysp_type*). But this probably doesn't matter, since
this class is usually short-lived anyway, it's not like the user will
have millions of these saved somewhere.

Change-Id: I01264c039110f87f88f5e92dea8ee3382f9169e2