Prefer protected members over explicit friends
Let's just "bring in" the SchemaNode constructor using protected access.
That's what the protected access is for, so let's save a few `friend`
statements here and there.

One still needs to befriend to parent class to the derived classes
because of the `asXYZ()` methods which need to invoke protected
constructors of the derived classes. For some reason GCC doesn't need
that, but clang does, and I think that clang is actually correct in this
one.

In the end I've left the access of the constructor in the derived
classes as `private` because I think these are not meant to be inherited
from.

Change-Id: I1557963cab160f7e2575caee2b2f9d80e2041fd3