Fix tree not freeing if overwriting with operator=
operator= only worked if both operand were from the same tree (same
m_refs). However, if they were different and the left operand was the
last ref to a tree, then the tree would leak, because up until now, the
tree was freed only in the destructor.

Side note: this bug only applies to operator= and not the copy
constructor, because ctors don't work with an existing object.

Change-Id: If55d8a07ee4baff90eaed88ac4b005f05feadaf2