Add minimal support for anydata with a lyd_node
The new DataNodeAny class can currently represent an anydata node with a
lyd_node (DataNode) inside it. This class will eventually be used for
all kinds of "anydata data" and also anyxml. But for now only a small
portion of all of that is supported.

Because of automatic memory management, it isn't simple to implement
value retrieval within DataNodeAny. That's why there is a method called
releaseValue() that disconnects the contained value from the DataNodeAny
and returns it. The user can then use the value independently, although
they can't retrieve it more than once. At the end, both are freed
separately.

Change-Id: Ia927a76b91ab456e868a9df2a92bbc6b6401d86f