Rework validateAll function
`lyd_validate_all` can indeed change the `node` argument to a nullptr.
This means that DataNode::validateAll could possibly leave the DataNode
instance in an invalid state. In libyang-cpp a nullptr of type lyd_node*
as encoded as an std::optional<DataNode> with a std::nullopt inside.
However, the class cannot morph itself into an optional, so validateAll
must a be free function. The argument is a non-const reference to an
optional data node. This means that we have a way of setting nullopt. We
also force the user to supply an std::optional argument.

Change-Id: Ib413db0db0f242014d97d4b02ebe1f3c4b0f39b2