Rework memory management
Working with low-level operations like unlink and insert (child, after,
before) requires having an algorithm that resolves memory management.
Technically, the only thing that needs to be done is to update the
m_refs field inside a group of siblings. Because this algorithm will be
the same for all of the operations, I have refactored it outside the
unlink method.

The algorithm is reworked:
1) Before, only one node could be updated. This made sense for unlink,
   because the new tree (created by unlink) would only consist of one
   parent and its children. However, there are algorithms, that work
   with siblings and not just one node (for example unlinkWithSiblings).
   Because of that, the algorithm has to be updated to be able to update
   the refcounters of many siblings.
2) Because of 1), it might be possible that ALL siblings from the
   original tree are now in the new tree. The algorithm needs a handle
   (lyd_node*) to the old tree. If it detects that all candidates (that
   is, all siblings) are from the same tree, then nothing is freed. This
   happens when the `nodes` vector consists of all sibling nodes.

The actual C-api call is made via a callback, because the algorithm does
stuff even after the call.

Change-Id: I55f4d1d2bb8459e581332ff3e1bcdf50abd53986