Refactor memory management from unlink
Working with low-level operations like unlink and insert (child, after,
before), requires having an algorithm that resolves memory management.
Technically, the only thing that needs to be is to update the m_refs
field inside DataNode. Because this algorithm will be the same for all
of the operations, I have refactored it outside the unlink method.

The actual C-api call is made via a callback, because the algorithm does
stuff even after the call.

Change-Id: I55f4d1d2bb8459e581332ff3e1bcdf50abd53986