handleLyTreeOperation: Use invalidateIterators
This is how it's done in all of the other collections. This is not a
bugfix.

The previous code inside Iterator<NodeType, ITER_TYPE>::throwIfInvalid()
had to also check for the m_valid flag inside the collection. Now that
we use invalidateIterators, instead of just setting the flag directly,
we don't have to check for that.

Change-Id: Idd11c324a43f6d871b6fbb75b95f52ced11b2753