Add Context::findXPath
Also add support for SchemaNode Set.

The advantage of findXPath over findPath is that you can use more
generic paths, similar to "schema paths" in libyang1.
Example: You have a path to a list and you want a SchemaNode of that but
you have no keys. findPath won't work, but findXPath will.

Change-Id: I21f7155a8883309da6936c0b8223632869978163