Rework unmanaged DataNodes
An unmanaged data node is characterized by that it's m_refs field is
nullptr. This means that all unmanaged data nodes have no access to the
libyang context. However, ALL data nodes have a context, so this patch
splits the internal_refcount struct and makes the context field
non-optional.

The reason for this change is that I need to use newMeta in opaque
nodes, which did not hold the context.

Change-Id: I86b89ddf7f892c77c1cba3d8ad9a5902d856858f