DataNode: Allowing attaching a custom context pointer
In sysrepo, a lyd_node can depend on a sr_data_t struct. This struct
contains the sysrepo connection which in turn contains the libyang
context. Freeing this sr_data_t can also cause the libyang context to be
freed. This means that if you wrap a lyd_node from the sr_data_t, the
created DataNode instance depends on the sr_data_t struct. The sr_data_t
struct must not be deleted before the lyd_node is deleted. DataNode must
somehow store sr_data_t struct.

To solve this, I added a custom shared_ptr that can contain additional
data supplied by the of this library.

Change-Id: I9405266ccbf0effbe5119360112bbf3d0e94e4cc