Rework bits
In libyang, the value of a `bits` leaf is represented by an array of
pointers to `bit` schema definitions. Because I was implementing schema
and data functionality separately, it happened that I created two
bit-representing structs that were almost the same. However, to decrease
confusion, I kept the Value-Bit struct (after discussion with Honza).

The only difference between the two structs is that the newer one is
safer, because it also held a reference to the context (since the `name`
field is a string_view and it is connected to the lifetime of the
context). My first thought was to use this safer struct, however, then
this struct could not be constructed in the tests. So, instead, I made
the `name` field a string and removed the dependency on the context.

This change should probably have been two changes, one to change
string_view to string and another to remove one of the structs, but oh
well.

Change-Id: I6ca30f485cc196eac93ee1ff645f2c4214619e28