Get rid of the String class
The purpose of libyang::String was to save a copy when converting to
std::string libyang-allocated const char*s. However, the only point
where this was actually useful was for printing. The users of the
library would usually convert the libyang::String to std::string anyway,
because they want to change it somehow (e.g. concatenate with something
else). This is true especially for paths, which aren't very long
(usually something like 100 at most?).

libyang::String was also used for printing data trees. They could for
example be sent over something like a NETCONF or RESTCONF. Although it
would probably be possible to use the underlying const char*, I suspect
it isn't very intuitive, and users will convert to std::string anyway.

Change-Id: I09aff4655087eeb910c1b132e1cde92fa4e5d566