Use std::string instead of const char*
The reason for using const char* was that the C++ API should have the
least overhead against the C API. However, in C++ you almost always
already have an std::string and a lot of times, you needed to create
temporaries for the arguments of libyang-cpp methods.

This patch changes all the arguments to std::string. In the future, we
can add const char* overloads if needed.

As for performance: the test suite isn't noticeable slower. Also, old
libyang C++ did it the same way and there were no issues.

Change-Id: I01cd52463bad9b65516ebadab6e7ad12cb31dc55