add a comparison operator for Identity
This took me a while. After a detour to the <=> operator,
strong_ordering, weak_ordering and some thinking about rejecting
cross-context comparisons, I ended up with a named operator. It is far
from elegant (I would have really preferred the multi-way semantics of a
strong_ordering, for example, but then that's not a thing that is
directly usable as a template argument for containers' element
comparisons).

Change-Id: Idfbd8b1d0bce62e57d5e7bc88b890edb8da19ee4