Start exporting just the "public" symbols
The main motivation is MSVC which doesn't export anything by default,
but hey, it's a nice feature even on Linux. Let's export only those
symbols which are really needed. GCC and clang have different
diagnostics when it comes to template classes, which is interesting, but
I have a nice #pragma hammer that I'm not afraid to use. I also had my
share of fun trying to figure out how to get GCC inherit the visibility
of an implicitly instantiated template method from the enclosing class.
I failed, which means that there's an explicit instantiation with an
appropriate attribute tag now.

And this is how it looks when something that is not exported gets used
from a test, for example:

 libyang-cpp/tests/unsafe.cpp:88: undefined reference to `libyang::Set<libyang::DataNode>::begin() const'

Keep the `friend` declaration in sync with the actual forward, otherwise
MSVC yells due to C4273 (inconsistent dll linkage).

Change-Id: I8d7962dc453fb6b7c8cd3707bbe6b61a8d3b021a