adapt to anydata handling changes
Our test was buggy because some of these strings were not syntaticaly
valid, which was definitely a bug. However, upstream libyang now
provides a data node instead of the original, raw strings, and there
appears to be a bug in the way how XML namespaces (the `xmlns`
attribute) are handled, and a behavior change for JSON arrays.

In the meanwhile, let's unbreak the build.

Bug: https://github.com/CESNET/libyang/issues/1958
Change-Id: I624d41c42f7c5204cc8cc6cacdbb8a582f9ad947