tests: working with anyxml
Upstream has changed the way how `anyxml` and `anydata` are handled, and
`anydata` is now turned into a tree of parsed data nodes. Our previous
versions of the test suite checked that when feeding XML (or JSON) in,
we can extract the original content in the same format. This now holds
for the `anyxml` data type, so let's add that one to the test suite.

Change-Id: If31edc2327d988090537844fb892f708fe7b8de6