refactor: less repetition when copying type constraint error info
C++ doesn't really do mixins that well (I think that type inheritance
would be evil in this context), so we have a bunch of type struct
members which need initialization. Let's try to do a bit less
repetition.

The possible alternative is to encapsulate the description and the error
info into a common struct, and use that struct as a member.

Change-Id: Ifc1ae2482c859c63b88576b84978b595d45a1974