anyxml: prevent use-after-free with JSON or XML raw data
I'm a bit torn on this change. Because the XML and JSON data structures
were written as very efficient thin wrappers that used a string_view,
there was a risk of a heap-use-after-free when the enclosing DataNode
went out of scope. Prevent that by using plain old strings.

Ultimately, I think this is a good change because:

- the only (reasonable) way of actually obtaining this AnydataValue
  instance is through DataNodeAny::releaseValue(),
- DataNodeAny is kinda useless without actually calling this function,
- we absolutely do need the current memory management of C++ wrappers,
- the only way of having this memory management in case of a DataNode
  enclosed in a DataNodeAny is calling an explicit
  release/detach/unlink/whatever, even though it modifies the original
  in place.

The (big) bonus is that the safety of AnydataValue no longer depends
on what was enclosed. It can still only be called once for DataNode and
N times for anything else, which is a bit confusing -- but at least it
won't crash. (I actually bumped into this behavior when refactoring the
test suite.)

The downside is that whenever the C++ bindings construct the XML/JSON
representation of `anyxml`/`anydata`, there will be some memory copying.

Change-Id: If40d26edfafd613d07592af657102c3c4384e087