build: check that all headers are self-contained
It's a new feature available since CMake 3.24. We cannot use that one
unconditionally, not even on Fedora 38, because buildroot uses 3.22 (and
that version doesn't have `FILE_SET`s, either). And it seems that
there's no easy way to make that target a part of `make all` because
adding either an explicit dependency, or setting the `EXCLUDE_FROM_ALL`
property doesn't work. Yay, cmake.

Change-Id: I9a14d9855e9e45caf94492e5c789798df0de1427