Avoid having to forward-declare std::variant<T>
The recent patches which added support for reporting numeric
ranges/limits of values were hit by a risk of a circular #include loop;
the std::variant that we use for reporting literal values (`Value`) was
needed from inside the `Type.hpp`. Simply forward-declaring the `Value`
broke the build on MSVC.

Since the only part that is needed is actually the `Identity` class, and
since that class is *not* something that's specific to types, but it's a
module-level statement, fix stuff by moving the `Identity` into
`Module.hpp`.

Change-Id: I5af3ed43ed9097094b7acd63672211afe0574967
Fixes: 7565fe3 ("refactor: forward-declare the Value variant")
Fixes: 5861137 ("schema: accessing numeric data limits")