Fix the import/export mess
Turns out that this symbol is actually needed for all builds because
it's used in the test suite already. Besides, it's meant as a public
function, so it really has to be made available to everybody.

The catch was that GCC- and clang-specific docs that I read suggested
that the proper place is after the function signature, as the very last
part, whereas MSVC-specific instructions pointed out that it belongs
right after the type definition. In the end the proper place is as the
very first thing in that statement, apparently.

Fixes: 510d5ac ("Fix symbol export/visibility")
Change-Id: I2a6017eaca61c0bc1a536cdd877b3404373a70ec