CI: sync the C++ dependencies
Unfortunately, we cannot depend on "new enough" doctest in CMake because
some external builds (TIP's Python bindings in particular) require
building on an old Ubuntu 20.04 which only has 2.3.6 (but because it's
on GitHub actions, it has a new enough compiler).

Let's try to replicate something like that here as well. On everything
but the GCC build, use the new doctest. On the most boring GCC build,
use the oldest doctest that we still care about.

Change-Id: I0732563cebed5e2e1d45affbdd5a1d788a9e4287