Fix Windows build due to std::filesystem::path::value_type and wchar_t
On windows, FS paths are built around a `wchar_t`, which leads to all
the possible joy that one could possibly have. Fortunately, the STL has
some converting functions. The only catch is that there are two path
formats, the "generic" one (same as on POSIX), and the native one. The
generic one on Windows uses forward slahses, whereas the native one on
Windows has backslashes. I think that the correct one to use here is the
native one because it's essentially garbage in, garbage out. AFAIK.

Change-Id: I2004f83014e02c6316048985be247c10bcc4cfe7