Add a wrapper for lyd_new_opaq()
In a recent release, upstream sysrepo added a function for discarding
previously pushed ops changes within a regular sr_apply_changes()
transaction. That's super nifty, but in a lot of our projects it's still
a bit messy to call a special function. It would be much more convenient
to just work with a single libyang in-memory tree, pushing any changes
as required.

It turns out that all that the `sr_discard_items()` function is doing is
just that, and that it's supported to create the required libyang data
nodes manually.

The libyang API is a bit messy, there are two functions, one for pushing
JSON data, others for XML data. Sysrepo calls the one for JSON even
though the data are *not* JSON in any way, so we're adding a wrapper for
the JSON thing. The actual usage looks funny (see the test), but the
resulting data tree appears to be the same as prepared by sysrepo
proper.

Change-Id: Ic7de1b0d1951c8bec6da2b36c52040d06dd60a93