Fix nullptr dereference when fetching module extensions
In commit fb60895 ("Allow getting module extensions") we implemented
getting extensions from YANG modules. However, in case the module was
not implemented, then the could crash on nullptr access. The libyang
docs for `lys_module` clearly say that the `compiled` field is
available only for implemented modules.

I have also added a new YANG module that imports ietf-restconf model.
This is for the tests only; we need to import ietf-restconf only, not
implement it to test the exceptions.

Fixes: fb60895835047d56c5fff4cfff42a46c92338ee2
Change-Id: I766e1c1267b304256049557bd4b5f3e056171d13