schema: Add support for `when` statement
While the RFC 7950 says that only some schema nodes support the `when`
statement, libyang provides access to a "compiled schema". In that
representation, `grouping` and `uses` are already resolved, which means
that even an `action` node might have an "effective" value of `when`.

Similarly, some "schema-only nodes", such as the `input`, `output`,
`case` and `choice` cannot have a `when` restriction. Both the C code
and the C++ wrapper are implemented in such a way that calling
`SchemaNode::when()` should be safe. Unfortunately, this is currently
untestable because there's no way to get such a `libayng::SchemaNode` in
the first place because these cannot be retrieved through the
`lys_find_path`.

Change-Id: I0ad952a592750a6b7ff2d563b0216148dbe93e0b
Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>