ABI break: instance-identifier: always preserve the original path
YANG defines the instance-identifier type to be essentially an XPath
that might point to an instance of a data tree node, or, in case the
`require-instance` is set to false in the schema, it might not point
anywhere at all.

In the previous version, an instance-identifier's value (i.e., our
`Value` variant) was always resolved into an instance of a DataNode.
That meant that in case of a value which does not point anywhere, the
original textual value of that XPath was lost, and the only way of
retrieving it would be to keep the original DataNodeTerm and to ask for
its string value. That's not pretty.

Solve that by introducing a "proxy" class which tracks both the XPath in
the text form, and the resolved node. We have to track both because the
C library uses an opaque structure for the "resolvable path to the
node", and there's a special C function for obtaining a lyd_node.
Furthermore, the std::optional<DataNode> cannot be used directly because
optional<T> won't work on forward-declared Ts. Well, std::any to the
rescue :).

This is obviously an API- and ABI-incompatible change.

Change-Id: I57b60db4a92d08e20563917e8a94ba4e2104e741