user-defined literals: fix working with very edge values
Apparently, this is a certain corner-case in the standard. These
user-defined literals are apparently always `unsigned long long`, the
minus is unary, and that means that there are special ways on how one
can write "the lowest negative number ever", and a magic macro is one of
these.

Then this requires some special casting, otherwise static_assert freaks
out about overflows and range checks. Yay for tests and for different
compilers.

Change-Id: I2129b993430338efdd49455be99f35586b0f52b3