fix DataNode::insertChild on a parent-less node with siblings
When a node which doesn't have any parent, but which has some following
siblings is getting inserted, the C library (by design) inserts
all if its siblings as well. This was not working well in the C++
wrapper, and the "thought to be orphaned part" of the original tree would
point to one of the siblings of the just-inserted node. As a result, the
C library knew that the siblings and the affected node belong to the
same forest (and are children of the same parent), while
libyang::DataNode would think that the siblings are managed by a
different internal_refcount.

The simplest example where this was crashing was a very simple case
where there was a single C++-level reference to node A (which has a
sibling B). When A was inserted somewhere, B was assumed to be
unreachable from the C++ wrapper, and its memory would be freed -- but
at that time the C library already knew that B is (still) a sibling
that's reachable from A, and freeing that one would release both. Hence,
a heap-use-after-free.

Fix this by some refactoring which ensures that handleLyTreeOperation()
(i.e., a helper for updating the C++'s internal_refcount) is always
called with a single node, and a flag which indicates whether the
operation affects all following-siblings. I used a second template
parameter because I didn't want to clutter up the public namespace by
some internal enum.

The explicit calls to `unlink` in the "positioned inserts" (which are
documented at the C level to be only used on single nodes and on
user-ordered lists) are removed to maintain parity with how the C
library works.

Change-Id: I86424c7a768f6b599f0709c2585230cbf68ed7eb