Rename OutputNodes enum to InputOutputNodes
libyang's lys_find_path() last parameter is ly_bool output. The
documentation states that:

  Search operation output instead of input.

I.e., it means that for e.g. the RPC nodes find_path can search for
input XOR output nodes. It can not find both types.
Libyang needs this because input and output nodes may have the same name
so the caller must decide which nodes he wants to consider.

This wrapper took this boolean and wrapped it in an enum OutputNodes
with possible values YES and NO. This corresponds almost 1:1 to the
ly_bool parameter in libyang.

The SchemaNode::findPath documentation does not mention the
parameter at all. The docs of the OutputNodes says:

  Controls whether output nodes should be considered when using findPath.

I find this very misleading. At first I did not know about the reason
why the output enum parameter is there and I thought that this only
makes findPath consider both input and output nodes. However, it turns
out that it consider input nodes only (when the value is NO) or output
nodes only (when the value is YES).

Although our current code corresponds to C libyang, I think we can take
the liberty and make this enum and its values more clear.
This patch renames the enum and its values and also improves the docs.

The change, however, changes the ABI.

Change-Id: Ib5de1c5d36b4448d7b3e3e62bd106373cd057182