module: add getter to get module namespace
When constructing RESTCONF RPC replies in XML format one needs to know
the module namespace.

In 2e320f1 ("Wrap `lyd_new_opaq2` and test RESTCONF-like RPC output") a
test for constructing a RESTCONF reply was added but the namespaces were
hardcoded.
In real application we need to retrieve the namespace. In C, it is
stored in lys_module::ns field but we did not have a getter for that.

This commit also modified the test for RESTCONF reply; it now retrieves
the module namespace from the node. The test should now look more
realistic.

Change-Id: Ida95dc819b968e3446e6910be78ef7aebc69677d