modernize: simplify unique_ptr's deleter
In the past we would use something "complex", like passing a C function
pointer which actually stores that function pointer in the type's
runtime memory, and dereferences the stored pointer-to-function at
destruction time. That's inefficient (whopping eight bytes, oh noes!),
so we want to "hardcode" the C function that's called into the object
type itself. Previously, this would require a named function, but since
C++20 which support lambdas in unevaluated contexts (whatever it is, but
yay), we can do better -- just use a decltype(lambda), and profit.

Looking at the diffstat, I'm not sure how much of a simplification this
is, but at least I've removed some duplicate code, and we're saving
eight bytes for (some of) these unique_ptr instances.

Change-Id: Ida14d29e141586cc29833836a692c839418dd355