schema printing
This is pretty straightforward; the only interesting part is a C++
wrapper for `ly_out_new_clb()` which just writes to a `std::string`. IAt
first I wanted to use `ly_out_new_mem`, but that one uses direct
reallocation, and this one is much simpler.

I was wondering if there's any value in swallowing exceptions when
inside a C callback, but in the end I decided that this is not wroth the
effort. Doing it "properly" would mean also testing for failures, and
that's either some completely custom CB, or a fault injection framework.
This evening was too short for either of these.

Change-Id: I114ce664caba14c4a1f2892beef750ad3646f1bb