Context::getModule *requires* a revision
In the C libyang, using a NULL revision actually means "get me that
version of a module if it exists", it's not a wildcard. The default
value in the C++ API however strongly suggested that this is actually a
shortcut for "the latest version of that module no matter if it's
implemented".

Fix that by always requiring that argument.

Change-Id: I23c477399a42d627623adea5b5484b550a6972bf