refactor: reusable conversion of std::vector<std::string> into a `const char **`
Also, don't bother with unique_ptr, just use a vector for this. This
version with std::vector is, IMHO, more readable.

This function might be unsafe if the original vector which holds these
strings goes out of scope before these pointers are consumed, but that's
equivalent to the previous version. As long as it's working on an
umodified parameter of the enclosing function, it's safe.

Change-Id: Id54c37add6f75b25491df5442756e435c91d419f